requirements.txt
aiohttp==3.8.3
attr==0.3.2
beautifulsoup4==4.11.1
brotli==1.0.9
ConfigParser==5.3.0
cryptography==3.4.8
Cython==0.29.32
discord.py==2.1.0
dl==0.1.0
docutils==0.19
HTMLParser==0.0.2
importlib_metadata==4.6.4
ipython==8.7.0
ipywidgets==8.0.2
Jinja2==3.1.2
keyring==23.5.0
lockfile==0.12.2
lxml==4.9.1
matplotlib==3.6.2
numpy==1.23.5
ordereddict==1.1
Pillow==9.3.0
player==0.6.1
protobuf==4.21.9
pydub==0.25.1
pyOpenSSL==22.1.0
python-dotenv==0.21.0
railroad==0.5.0
scipy==1.9.3
soundfile==0.11.0
Sphinx==5.3.0
toml==0.10.2
tornado==6.2
zipp==1.0.0
